fix task108

+++ src/Chapter1/SimpleLoops.java
@@ -41,9 +41,9 @@ public class SimpleLoops {
 
     public static int task108(int n) {
         if (Math.pow(2, (Math.log(n) / Math.log(2))) == n) {
-            return (Math.log(n) / Math.log(2));
+            return (int)((Math.log(n) / Math.log(2)));
         } else {
-            return (Math.log(n) / Math.log(2)) + 1;
+            return (int)((Math.log(n) / Math.log(2)) + 1);
         }
     }
     public static boolean task88a(int num){
